+++ provider/acc_project_test.go
@@ -191,3 +191,393 @@ func TestRecreateProjectIfNotFound(t *testing.T) {
 			})
 	})
 }
+
+func TestPrimaryCompute(t *testing.T) {
+	if os.Getenv("TF_ACC") != "1" {
+		t.Skip("TF_ACC must be set to 1")
+	}
+
+	client, err := neon.NewClient(neon.Config{Key: os.Getenv("NEON_API_KEY")})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	projectNamePrefix += "primaryCompute-"
+
+	t.Cleanup(func() {
+		resp, _ := client.ListProjects(nil, nil, &projectNamePrefix, nil, nil)
+		for _, project := range resp.Projects {
+			_, _ = client.DeleteProject(project.ID)
+		}
+	})
+
+	var newProjectName = func() string {
+		return projectNamePrefix + strconv.FormatInt(time.Now().UnixMilli(), 10)
+	}
+
+	t.Run("shall create a new project with different compute configs and the default compute's configs",
+		func(t *testing.T) {
+			projectName := newProjectName()
+			resource.Test(
+				t, resource.TestCase{
+					ProviderFactories: map[string]func() (*schema.Provider, error){
+						"neon": func() (*schema.Provider, error) {
+							return newAccTest(), nil
+						},
+					},
+					Steps: []resource.TestStep{
+						{
+							Config: fmt.Sprintf(`resource "neon_project" "this" {
+		name = "%s"
+		autoscaling_limit_min_cu = 0.25
+		autoscaling_limit_max_cu = 1
+		suspend_timeout_seconds  = 300
+		
+		primary_compute {
+			autoscaling_limit_min_cu = 0.5
+			autoscaling_limit_max_cu = 2
+			suspend_timeout_seconds  = -1
+		}
+}
+`, projectName),
+							Check: resource.ComposeTestCheckFunc(
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"name", projectName,
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_min_cu", "0.25",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_max_cu", "1",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"suspend_timeout_seconds", "300",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_min_cu", "0.5",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_max_cu", "2",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.suspend_timeout_seconds", "-1",
+								),
+								func(_ *terraform.State) error {
+									got, err := readProjectInfo(client, projectName)
+									if err != nil {
+										return err
+									}
+									assert.Equalf(t, int64(300),
+										int64(*got.DefaultEndpointSettings.SuspendTimeoutSeconds),
+										"expected compute defaults: suspend_timeout_seconds")
+									assert.Equalf(t, float64(0.25),
+										float64(*got.DefaultEndpointSettings.AutoscalingLimitMinCu),
+										"expected compute defaults: autoscaling_limit_min_cu")
+									assert.Equalf(t, float64(1),
+										float64(*got.DefaultEndpointSettings.AutoscalingLimitMaxCu),
+										"expected compute defaults: autoscaling_limit_max_cu")
+
+									e, err := client.ListProjectEndpoints(got.ID)
+									if err != nil {
+										return err
+									}
+									defaultEndpoint := e.Endpoints[0]
+									assert.Equalf(t, int64(-1),
+										int64(defaultEndpoint.SuspendTimeoutSeconds),
+										"expected default compute's suspend_timeout_seconds")
+									assert.Equalf(t, float64(0.5),
+										float64(defaultEndpoint.AutoscalingLimitMinCu),
+										"expected default compute's autoscaling_limit_min_cu")
+									assert.Equalf(t, float64(2),
+										float64(defaultEndpoint.AutoscalingLimitMaxCu),
+										"expected default compute's autoscaling_limit_max_cu")
+
+									return nil
+								},
+							),
+						},
+					},
+				})
+		})
+
+	t.Run("shall update default compute's configs w/o affecting project's compute configs",
+		func(t *testing.T) {
+			projectName := newProjectName()
+			resource.Test(
+				t, resource.TestCase{
+					ProviderFactories: map[string]func() (*schema.Provider, error){
+						"neon": func() (*schema.Provider, error) {
+							return newAccTest(), nil
+						},
+					},
+					Steps: []resource.TestStep{
+						{
+							Config: fmt.Sprintf(`resource "neon_project" "this" {
+		name = "%s"
+		autoscaling_limit_min_cu = 0.25
+		autoscaling_limit_max_cu = 1
+		suspend_timeout_seconds  = 300
+		
+		primary_compute {
+			autoscaling_limit_min_cu = 0.5
+			autoscaling_limit_max_cu = 2
+			suspend_timeout_seconds  = -1
+		}
+}
+`, projectName),
+							Check: resource.ComposeTestCheckFunc(
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"name", projectName,
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_min_cu", "0.25",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_max_cu", "1",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"suspend_timeout_seconds", "300",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_min_cu", "0.5",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_max_cu", "2",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.suspend_timeout_seconds", "-1",
+								),
+							),
+						},
+						{
+							Config: fmt.Sprintf(`resource "neon_project" "this" {
+		name = "%s"
+		autoscaling_limit_min_cu = 0.25
+		autoscaling_limit_max_cu = 1
+		suspend_timeout_seconds  = 300
+		
+		primary_compute {
+			autoscaling_limit_min_cu = 1
+			autoscaling_limit_max_cu = 4
+			suspend_timeout_seconds  = 1200
+		}
+}
+`, projectName),
+							Check: resource.ComposeTestCheckFunc(
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"name", projectName,
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_min_cu", "0.25",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_max_cu", "1",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"suspend_timeout_seconds", "300",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_min_cu", "1",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_max_cu", "4",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.suspend_timeout_seconds", "1200",
+								),
+								func(_ *terraform.State) error {
+									got, err := readProjectInfo(client, projectName)
+									if err != nil {
+										return err
+									}
+									assert.Equalf(t, int64(300),
+										int64(*got.DefaultEndpointSettings.SuspendTimeoutSeconds),
+										"expected compute defaults: suspend_timeout_seconds")
+									assert.Equalf(t, float64(0.25),
+										float64(*got.DefaultEndpointSettings.AutoscalingLimitMinCu),
+										"expected compute defaults: autoscaling_limit_min_cu")
+									assert.Equalf(t, float64(1),
+										float64(*got.DefaultEndpointSettings.AutoscalingLimitMaxCu),
+										"expected compute defaults: autoscaling_limit_max_cu")
+
+									e, err := client.ListProjectEndpoints(got.ID)
+									if err != nil {
+										return err
+									}
+									defaultEndpoint := e.Endpoints[0]
+									assert.Equalf(t, int64(1200),
+										int64(defaultEndpoint.SuspendTimeoutSeconds),
+										"expected default compute's suspend_timeout_seconds")
+									assert.Equalf(t, float64(1),
+										float64(defaultEndpoint.AutoscalingLimitMinCu),
+										"expected default compute's autoscaling_limit_min_cu")
+									assert.Equalf(t, float64(4),
+										float64(defaultEndpoint.AutoscalingLimitMaxCu),
+										"expected default compute's autoscaling_limit_max_cu")
+
+									return nil
+								},
+							),
+						},
+					},
+				})
+		})
+
+	t.Run("shall update project's compute configs w/o affecting default compute's configs",
+		func(t *testing.T) {
+			projectName := newProjectName()
+			resource.Test(
+				t, resource.TestCase{
+					ProviderFactories: map[string]func() (*schema.Provider, error){
+						"neon": func() (*schema.Provider, error) {
+							return newAccTest(), nil
+						},
+					},
+					Steps: []resource.TestStep{
+						{
+							Config: fmt.Sprintf(`resource "neon_project" "this" {
+		name = "%s"
+		autoscaling_limit_min_cu = 0.25
+		autoscaling_limit_max_cu = 1
+		suspend_timeout_seconds  = 300
+		
+		primary_compute {
+			autoscaling_limit_min_cu = 0.5
+			autoscaling_limit_max_cu = 2
+			suspend_timeout_seconds  = -1
+		}
+}
+`, projectName),
+							Check: resource.ComposeTestCheckFunc(
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"name", projectName,
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_min_cu", "0.25",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_max_cu", "1",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"suspend_timeout_seconds", "300",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_min_cu", "0.5",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_max_cu", "2",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.suspend_timeout_seconds", "-1",
+								),
+							),
+						},
+						{
+							Config: fmt.Sprintf(`resource "neon_project" "this" {
+		name = "%s"
+		autoscaling_limit_min_cu = 0.5
+		autoscaling_limit_max_cu = 2
+		suspend_timeout_seconds  = 600
+		
+		primary_compute {
+			autoscaling_limit_min_cu = 0.5
+			autoscaling_limit_max_cu = 2
+			suspend_timeout_seconds  = -1
+		}
+}
+`, projectName),
+							Check: resource.ComposeTestCheckFunc(
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"name", projectName,
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_min_cu", "0.5",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"autoscaling_limit_max_cu", "2",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"suspend_timeout_seconds", "600",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_min_cu", "0.5",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.autoscaling_limit_max_cu", "2",
+								),
+								resource.TestCheckResourceAttr(
+									"neon_project.this",
+									"primary_compute.0.suspend_timeout_seconds", "-1",
+								),
+								func(_ *terraform.State) error {
+									got, err := readProjectInfo(client, projectName)
+									if err != nil {
+										return err
+									}
+									assert.Equalf(t, int64(600),
+										int64(*got.DefaultEndpointSettings.SuspendTimeoutSeconds),
+										"expected compute defaults: suspend_timeout_seconds")
+									assert.Equalf(t, float64(0.5),
+										float64(*got.DefaultEndpointSettings.AutoscalingLimitMinCu),
+										"expected compute defaults: autoscaling_limit_min_cu")
+									assert.Equalf(t, float64(2),
+										float64(*got.DefaultEndpointSettings.AutoscalingLimitMaxCu),
+										"expected compute defaults: autoscaling_limit_max_cu")
+
+									e, err := client.ListProjectEndpoints(got.ID)
+									if err != nil {
+										return err
+									}
+									defaultEndpoint := e.Endpoints[0]
+									assert.Equalf(t, int64(-1),
+										int64(defaultEndpoint.SuspendTimeoutSeconds),
+										"expected default compute's suspend_timeout_seconds")
+									assert.Equalf(t, float64(0.5),
+										float64(defaultEndpoint.AutoscalingLimitMinCu),
+										"expected default compute's autoscaling_limit_min_cu")
+									assert.Equalf(t, float64(2),
+										float64(defaultEndpoint.AutoscalingLimitMaxCu),
+										"expected default compute's autoscaling_limit_max_cu")
+
+									return nil
+								},
+							),
+						},
+					},
+				})
+		})
+}